restructured SMC and Metropolis samplers, Metropolis cleanup
